backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-openai
python-dotenv
pydantic
